Added backend tweaks.

+++ src/BidCalculator/BidCalculator.Server/Core/AssociationFee.cs
@@ -1,6 +1,6 @@
 ﻿namespace BidCalculator.Server.Core;
 
-public class AssociationFee(double amount) : Fee(amount)
+public class AssociationFee(decimal amount) : Fee(amount)
 {
-	public override string Name => "Association";
+    public override string Name => "Association";
 }
