+++ src/BidCalculator/BidCalculator.Server/Core/BasicFee.cs
@@ -1,6 +1,6 @@
 ﻿namespace BidCalculator.Server.Core;
 
-public class BasicFee(double amount) : Fee(amount)
+public class BasicFee(decimal amount) : Fee(amount)
 {
-	public override string Name => "Basic";
+    public override string Name => "Basic";
 }
