+++ src/BidCalculator/BidCalculator.Server/Core/Fee.cs
@@ -1,7 +1,7 @@
 ﻿namespace BidCalculator.Server.Core;
 
-public abstract class Fee(double amount)
+public abstract class Fee(decimal amount)
 {
-	public abstract string Name { get; }
-	public double Amount { get; set; } = Math.Round(amount, 2);
+    public abstract string Name { get; }
+    public decimal Amount { get; set; } = Math.Round(amount, 2);
 }
