+++ src/BidCalculator/BidCalculator.Server/Core/VehicleCostAnalysis.cs
@@ -1,7 +1,7 @@
 ﻿namespace BidCalculator.Server.Core;
 
-public class VehicleCostAnalysis (double basePrice, IEnumerable<Fee> applicableFees)
+public class VehicleCostAnalysis(decimal basePrice, IEnumerable<Fee> applicableFees)
 {
-	public IEnumerable<Fee> Fees { get; } = applicableFees;
-	public double Total => Math.Round(basePrice + Fees.Sum(f => f.Amount), 2);
+    public IEnumerable<Fee> Fees { get; } = applicableFees;
+    public decimal Total => Math.Round(basePrice + Fees.Sum(f => f.Amount), 2);
 }
